Update README.textile
Noted that this branch should NOT be used as is. Code is incomplete and / or non linking in many places (on purpose) as it is being developed in my spare time, so I might only do a partial code one day, thus DO NOT USE this branch, use the master if you want to play with the fork.